+++ 12209388-Riska dwi apriliani/vue akhir/src/components/Blog/Card.vue
